Initial commit of OAuth2 API for the user manager, adds oauth2 clients (called services by the user manager) to the database, as well as service connections (session contexts in the database) and sessions.
By coming in through an oauth2 API users can get the user to interactively create a service connection (the initial grant of basic access to user info to a service the first time a user accesses some service) and then automatically create sessions and be directed back to the calling application.

Basic groundwork for client-side oauth2 logic (we register a callback service and will additionally require a custom user provider to be built so that we can perform the initial redirect of the user).