Swap WebQuery backend from Criteria (no longer the preferred route) to HQL (#29)
Convert the backend of WebQuery, maintaining almost complete backwards compatibility with the existing WebQuery syntax. The one break in backwards compatibility happens for composite primary keys (EmbeddedId annotation): the old code mapped these to top-level properties - e.g. given an EmbeddedId id with an Embeddable PK type of SomePrimaryKey(id, timestamp) as in the case with the EmbeddedPkEntity.java test case, the old code provided properties of "id" and "timestamp". Criteria+HQL APIs natively talk about these properites as "id.id" and "id.timestamp". The new WebQuery HQL back-end exposes these properties as "id:id" and "id:timestamp" (the reason for the colon instead of a dot is to allow a simple split-on-dot approach to be used).

This change also removes the caching of full dotted string paths -> path to properties which was performed in the old Criteria code, it did not appear to make a significant performance improvement. If the new code proves sluggish then it could be reintroduced with only minor changes.

* Initial working commit of WebQuery -> HQL conversion (from using Criteria previously). This is a very much in development commit with many unit tests not passing currently

* Fix DynamicQueryTest.testGetByUniquePropertyOnCollection (was using wrong code to determine if a relation referred to a collection)

* Fix HQL generation for passing in collections (as a result of calling HQBuilder.addIdInConstraint). HQL Query must be instructed that parameter is a collection rather than relying on it to detect this itself

* Identified an issue in the WebQuery->HQL code when using HSQLDB which was triggered by the test case for the logic we use to compensate for a SQL Server issue (querying large tables with many joins was resulting in SQL Server creating a massive temp table with all the joins, then filtering down to a small number of rows - which was killing performance for certain applications).
Rather than remoce the unit test I made the code work with HSQLDB too.
The WebQuery->HQL issue is because Hibernate converts the HQL rather directly to SQL and HSQLDB has a limitation about columns being referenced in ORDER BY if they are not being SELECTed.
The SQL Server workaround goes like this: instead of applying constraints and fetching back data simultaneously, we fetch back only the IDs of constrained entities, then in a separate query we fetch back the data associated with those entities.

The fix is to ensure that when an ID projection is being used in HQBuilder, we also SELECT the ORDER BY columns and then discard them as they're returned by the database.

* When the COUNT projection is used, don't include pagination/ordering parts of the query

* Add unit test that fails under HSQLDB for ordering by a lazy-joined child relationship - will probably require some amount of effort to solve this issue, but isn't a major issue because HSQLDB is only used to run unit tests

* Make the HSQLDB fix be configurable, and default to false. Enable the HSQLDB fix for all db unit tests.
N.B. does not fix unit test that selects data and orders by unselected data

* Fix the use of discriminators for WebQuery->HQL. Also enhance the subclass logic so that a query that references a subclass is then allowed to use the fields from that subclass (and so the query will fetch back instances of that subclass, not just of the generic superclass originally queried)

* Fix the use of composite primary keys in WebQuery->HQL logic. Previously these keys were mapped in as a simple column and then translated at the back-end to a dotted property name, changing it instead to be more explicit and to use colons (like with a collection:size property, but id:timestamp etc.). This is a breaking change, however no code actually uses this way of mapping primary keys within a WebQuery.

* Implement alias support in WebQuery->HQL

* Allow DateTime to be created from a stored Long

* Allow log4j logging of SQL statements as they are logged

* Various modifications:
 - SELECT all ORDER BY columns
 - Automatically set transform to ROOT_ENTITY if we're adding column SELECTs just to satisfy the DBMS
 - Reverse order of added columns, HQL appears to flip the order (so the last SELECTed column/entity appears first). This is necessary for ROOT_ENTITY transform to work
 - Instead of explicitly mentioning "mobj" everywhere, allow the use of "{alias}" like in Hibernate's Criteria SQL extensions.
 - Make the functionality that allows the definition of {alias} generic enough that HQLBuilder users can use it to create auto-expansions to prevent repetition in their query extensions (where they're adding custom HQL)
 - Allow specifying a custom HQL SELECT statement
 - Allow adding custom HQL ORDER BY statements
 - Allow adding custom HQL GROUP BY statements

* With recent change to SELECT all ORDER BY columns the test testOrderingByLazyAssociatedRelationThatIsNotSelectedDoesNotWorkWithHSQLDB now works against HSQLDB

* Improve unit test name and documentation for ordering by unselected column