This commit addresses an issue we've observed with the oauth2 flow for services, whereby a user clicking a link without having an active session with a client service can cause 2 oauth flows in parallel (may be because of double-clicking on a link) - specifically we saw 2 oauth2 flows start 40ms apart. They both returned from the oauth2 provider, but only one succeeded (because 2 sessions were created, so the browser used the most recent one). This then resulted in one link getting an error about an unexpected auth attempt, and the other session logging in and returning to the desired resource in the client service.
This commit changes the verification of the "state" function, such that on an internal error, we ignore the nonce value check and simply follow the redirect; the reason for this is that we will (and can, due to using an HTTP 303 See Other redirect) only issue a GET redirect, we don't allow re-running any POST operation automatically. This means that malicious users can't trick users into clicking a link that will result in a modification operation, only a GET operation (which they could trigger anyway via a direct malicious link). The only change is the value of the Referer header, but that header shouldn't (and isn't) used for security purposes.

Refactor OAuth2 Client Callback handling such that:
 - Scope for OAuth2 Flow includes the method (must be GET currently)
 - On an OAuth2 Client Callback processing failure, if a GET request, redirect the client browser to the redirectTo endpoint (this will restart the oauth2 auth flow if necessary, or if the user has already successfully logged in, it will simply return the desired resource).