Switch from log4j to logback (#40)
Switches the logging subsystem from log4j to logback, and adds the slf4j adapters for commons-logging and log4j so that existing code works as-is.

This change includes a converter for very simple log4j.properties files, so simple services can run without any changes. The converter coverts a single standard output pattern appender, the default level and your per-class/package levels.

Or you can use a logback xml (which is the recommended approach). To do this, simply set the following config:

logback-config=/logging.xml

And in tomcat/shared/lib/logging.xml (or elsewhere in your config structure):

<configuration>
	<appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
		<encoder class="ch.qos.logback.classic.encoder.PatternLayoutEncoder">
			<Pattern>%d{ISO8601} %5p %m%n</Pattern>
		</encoder>
	</appender>

	<logger name="some.package.name" level="INFO"/>

	<root level="WARN">
		<appender-ref ref="STDOUT"/>
	</root>
</configuration>

It is possible to use ${...} variable references in this logging.xml (just as with other .properties files in the framework). The variables will be resolved before the XML is passed to logback (otherwise you'd need some incantations to get a connection back to us for properties)